refactor: refactor some styles
